[FR #13] Create Command Classes && Command Visitor
Each highlighted user action corresponds to a certain `*Command` class.
This class implements the Visitor pattern. Further logic of
interaction with objects of `*Command` classes will be implemented
exactly in the Visitor class.

Further adding commands requires creating a new class that is
a descendant of `ACommand` and adding the appropriate method
to the CommandVisitor