[FR #12] Create Query Classes && Query Visitor
Each highlighted user query corresponds to a certain `*Query` class.
This class implements the Visitor pattern. Further logic of
interaction with objects of `*Query` classes will be implemented
exactly in the Visitor class.

Further adding queries requires creating a new class that is
a descendant of `AQuery` and adding the appropriate method
to the QueryVisitor.