[FR #14] Create abstract factory && TableGateways
As a result of the design, it was decided to interact with the DB
layer by using the TableGateway pattern. To control the belonging
of all created classes to a single family, the Abstract Factory
pattern was introduced.

TableGatewayFactory is an implementation of this pattern,
and is a Singleton object. Registration of factories is done
by calling its method `registerFactory`. A custom exception
class has been added for empty option errors.

To test Singleton's correctness, it was covered by unit tests.
To do this, the necessary dependencies were added to the build.sbt file.